chore: rewrite preview4 to js
