+++ src/previews/Preview4/Preview4.jsx
@@ -1,19 +1,17 @@
-import React, { FunctionComponent } from 'react';
-import { Colors, InPlayerIcon } from '@inplayer-org/inplayer-ui';
+// @flow
+import React from 'react';
 import { transparentize } from 'polished';
-import { withTheme } from "styled-components";
 
 // Images
-import previewImg from 'assets/ip-preview-premium.png';
-import restrictedAssetImg from 'assets/restricted-asset.png';
-
-// Types
-import { Branding } from './types';
+import previewImg from 'assets/images/ip-preview-premium.png';
+import restrictedAssetImg from 'assets/images/restricted-asset.png';
 
 // Components
+import { InPlayerIcon } from 'elements';
+import colors from 'config/colors';
 import StyledContainer from '../components/StyledContainer';
 import TextEditor from '../components/TextEditor';
-import { OverlayLabel, PreviewImage } from './styledComponents/SharedComponents';
+import { OverlayLabel, PreviewImage } from '../components/SharedComponents';
 import {
   StyledPreviewBox,
   ImageWrapper,
@@ -22,20 +20,21 @@ import {
   StyledTextWrapper,
   StyledIconWrapper,
   Icon,
-} from './styledComponents/Preview4';
+} from './styled';
 
-interface Props {
-  branding?: Branding;
-  assetCountrySetId?: number;
-  assetDomainRestrictions?: Array<any>;
-  displayBuyButton?: boolean;
-  previewNotAvailable?: boolean;
-  defaultAssePreviewDetails?: boolean;
-  isFullPreview?: boolean;
-  width?: string;
-  height?: string;
-  theme?: any;
-}
+// Types
+import { Branding } from '../types';
+
+type Props = {
+  branding?: Branding,
+  assetCountrySetId?: number,
+  assetDomainRestrictions?: Array<any>,
+  displayBuyButton?: boolean,
+  previewNotAvailable?: boolean,
+  isFullPreview?: boolean,
+  width?: number,
+  height?: number,
+};
 
 const removeTags = (str: string) =>
   str
@@ -43,15 +42,15 @@ const removeTags = (str: string) =>
     .replace(/ +/g, ' ')
     .trim();
 
-const Preview4: FunctionComponent<Props> = ({
+const Preview4 = ({
   branding: {
     inplayer_protected_label: protectedLabel = true,
     paywall_cover_photo: imageUrl = previewImg,
     preview_title: previewTitle = `<h1><strong>Asset Title</strong></h1>`,
     preview_description: previewDescription = `<p>Asset Description</p>`,
     preview_button_label: previewButtonLabel = 'Buy',
-    preview_buttons_bg_color: buttonBgColor = Colors.green,
-    preview_buttons_text_color: buttonTextColor = Colors.white,
+    preview_buttons_bg_color: buttonBgColor = colors.green,
+    preview_buttons_text_color: buttonTextColor = colors.white,
   } = {},
   assetCountrySetId,
   assetDomainRestrictions = [],
@@ -63,7 +62,6 @@ const Preview4: FunctionComponent<Props> = ({
 }: Props) => {
   const isRestrictedAsset = assetCountrySetId || assetDomainRestrictions.length !== 0;
   const assetPreviewImage = isRestrictedAsset ? restrictedAssetImg : imageUrl || previewImg;
-
   const previewTitleText = removeTags(previewTitle);
   const previewDescriptionText = removeTags(previewDescription);
   const widthNumber = width ? Number(width.slice(0, -1)) : 0;
@@ -96,7 +94,7 @@ const Preview4: FunctionComponent<Props> = ({
             <TextEditor
               value={previewTitle}
               displayToolbar={false}
-              textBackground={previewTitleText ? transparentize(0.3, Colors.black) : 'none'}
+              textBackground={previewTitleText ? transparentize(0.3, colors.black) : 'none'}
               textColor={buttonBgColor}
               isPadding
               readOnly
@@ -106,8 +104,8 @@ const Preview4: FunctionComponent<Props> = ({
             <TextEditor
               value={previewDescription}
               displayToolbar={false}
-              textBackground={previewDescriptionText ? transparentize(0.3, Colors.black) : 'none'}
-              textColor={Colors.white}
+              textBackground={previewDescriptionText ? transparentize(0.3, colors.black) : 'none'}
+              textColor={colors.white}
               lineHeight={1.5}
               isPadding
               readOnly
@@ -119,4 +117,15 @@ const Preview4: FunctionComponent<Props> = ({
   );
 };
 
-export default withTheme(Preview4);
+Preview4.defaultProps = {
+  branding: {},
+  assetCountrySetId: 0,
+  assetDomainRestrictions: [],
+  displayBuyButton: false,
+  previewNotAvailable: false,
+  isFullPreview: false,
+  width: 0,
+  height: 0,
+};
+
+export default Preview4;
